+++ src/realm/object-store/shared_realm.cpp
@@ -194,22 +194,24 @@ std::shared_ptr<SyncSession> Realm::sync_session() const
 
 sync::SubscriptionSet Realm::get_latest_subscription_set()
 {
-    // If there is a subscription store, then return the latest set
-    if (auto flx_sub_store = m_coordinator->sync_session()->get_flx_subscription_store()) {
-        return flx_sub_store->get_latest();
+    if (!m_config.sync_config || !m_config.sync_config->flx_sync_requested) {
+        throw std::runtime_error("Flexible sync is not enabled");
     }
-    // Otherwise, throw runtime_error
-    throw std::runtime_error("Flexible sync is not enabled");
+    // If there is a subscription store, then return the active set
+    auto flx_sub_store = m_coordinator->sync_session()->get_flx_subscription_store();
+    REALM_ASSERT(flx_sub_store);
+    return flx_sub_store->get_latest();
 }
 
 sync::SubscriptionSet Realm::get_active_subscription_set()
 {
-    // If there is a subscription store, then return the active set
-    if (auto flx_sub_store = m_coordinator->sync_session()->get_flx_subscription_store()) {
-        return flx_sub_store->get_active();
+    if (!m_config.sync_config || !m_config.sync_config->flx_sync_requested) {
+        throw std::runtime_error("Flexible sync is not enabled");
     }
-    // Otherwise, throw runtime_error
-    throw std::runtime_error("Flexible sync is not enabled");
+    // If there is a subscription store, then return the active set
+    auto flx_sub_store = m_coordinator->sync_session()->get_flx_subscription_store();
+    REALM_ASSERT(flx_sub_store);
+    return flx_sub_store->get_active();
 }
 #endif
 
@@ -1091,9 +1093,19 @@ void Realm::convert(const Config& config, bool merge_into_existing)
     verify_thread();
 
 #if REALM_ENABLE_SYNC
-    if (config.sync_config && config.sync_config->flx_sync_requested) {
-        throw std::logic_error("Realm cannot be converted if flexible sync is enabled");
+    auto src_is_flx_sync = m_config.sync_config && m_config.sync_config->flx_sync_requested;
+    auto dst_is_flx_sync = config.sync_config && config.sync_config->flx_sync_requested;
+    auto dst_is_pbs_sync = config.sync_config && !config.sync_config->flx_sync_requested;
+
+    if (dst_is_flx_sync && !src_is_flx_sync) {
+        throw std::logic_error(
+            "Realm cannot be converted to a flexible sync realm unless flexible sync is already enabled");
     }
+    if (dst_is_pbs_sync && src_is_flx_sync) {
+        throw std::logic_error(
+            "Realm cannot be converted from a flexible sync realm to a partition based sync realm");
+    }
+
 #endif
 
     if (merge_into_existing && util::File::exists(config.path)) {