+++ test/benchmark-util-network/main.cpp
@@ -3,6 +3,7 @@
 #include <thread>
 #include <iostream>
 
+#include <realm/status.hpp>
 #include <realm/sync/network/network.hpp>
 
 #include "../util/timer.hpp"
@@ -80,7 +81,7 @@ private:
         if (m_num_posts == 0)
             return;
         --m_num_posts;
-        m_service.post([=]() {
+        m_service.post([=](Status) {
             initiate();
         });
     }